backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-multipart
passlib[bcrypt]
pyjwt
requests
python-docx
python-pptx
python-dotenv
